pyproject.toml
[project]
name = "strands"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "strands-agents[gemini]>=1.26.0",
    "strands-agents-tools>=0.2.20",
]
